Update the platforms for Melissa
One dep of Melissa is not available on armv6l

Package ‘ghc-8.10.7’ in /nix/store/hjaixgpg8m0xafqixjdgrd4i9gyq6irh-source/pkgs/development/compilers/ghc/8.10.7.nix:344 is not supported on ‘armv6l-linux’, refusing to evaluate.

a bit tricky, but I added `ghc` as a dep and use its platforms.
but not as pretty as I wish