Update slurm platforms
one of the deps of slurms is not available on x86_64-darwin:

Package ‘rdma-core-37.1’ in /nix/store/hjaixgpg8m0xafqixjdgrd4i9gyq6irh-source/pkgs/os-specific/linux/rdma-core/default.nix:45 is not supported on ‘x86_64-darwin’, refusing to evaluate.

and one of the deps of slurms is not available on armv6l-linux:

Package ‘ghc-8.10.7’ in /nix/store/hjaixgpg8m0xafqixjdgrd4i9gyq6irh-source/pkgs/development/compilers/ghc/8.10.7.nix:344 is not supported on ‘armv6l-linux’, refusing to evaluate.

We thus take the intersection of the platforms